Separate failure detection from topic updates in statestore
As originally envisaged, the statestore was not intended to transmit GBs
of data to large clusters. Now that it is, some changes to the design
are required to improve stability and responsiveness.

One of the most acute problems is the way that the topic updates, which
may be extremely large, were responsible for conveying failure detection
information. The delay in sending topic heartbeats (which were often
blocked behind a queue of other updates) meant that large clusters
needed to set large timeouts for every subscriber, and to reduce the
frequency at which topic updates were sent.

This change adds a second heartbeat type to the statestore which is
responsibie only for coordinating failure information between the
subscriber and the statestore. These 'keep-alive' heartbeats are sent
much more frequently, as they have a tiny payload and do very little
work. The statestore now does not use the result of topic updates to
maintain its view of which subscribers are alive or dead, but only the
results of the keep-alive heartbeats.

This patch is a bit of a WIP, in that it's hard to be sure that these
changes are effective without testing on a large cluster with a large
catalog. One particular concern is whether keep-alive heartbeats will be
starved if there are many threads trying to send large catalog
heartbeats.

Once that's done, I think we can set more reasonable defaults
for the various configuration options and have a better shot at the
statestore working with a minimum of configuration out of the box.

Change-Id: Ia447e4ebefda890a5b810a213e97f00cca499989
Reviewed-on: http://gerrit.cloudera.org/4
Reviewed-by: Henry Robinson <henry@cloudera.com>