IMPALA-181: ORDER BY with Ordinals
In case of certain queries order by with ordinals would not work
properly. This is the case for all "select * " type of queries. Until
now, the ordinal substitution was based on the values from the select
list. However, these expression are not expanded in case of "*",
rather the list of result expressions and column lables is filled.

This patch simply changes the lookup of the expression from the select
list to the result list because only ordinals from the result can be
used as a sorting field.

Change-Id: Ie32d30375163d8ebc0f6c06a6fdec0a7c38c6b56
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/4808
Reviewed-by: Martin Grund <mgrund@cloudera.com>
Tested-by: jenkins