CDH-22383: Impala hangs when querying HBase tables with large number of columns
This commit fixes the issue where Impala hangs when querying an HBase
table with large (>500) number of columns. The issue was triggered by a
large memory allocation of a tuple buffer during the first GetNext call
of the HBase scanner that was causing an infinite loop where each iteration
was allocating a significant amount of memory. The fix is to dynamically
set the mem limit of a row batch based on the corresponding row size and to
dynamically set the maximum size of the tuple buffer so that it does not
exceed that limit.

Change-Id: Ia64f98b229772b50658af952fc641bf00f54f450
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/4871
Reviewed-by: Dimitris Tsirogiannis <dtsirogiannis@cloudera.com>
Tested-by: jenkins
(cherry picked from commit 87f382aa8fb44592d84070e2e8cd7aa0df3b3bd7)
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/4934
Tested-by: Juan Yu <jyu@cloudera.com>
(cherry picked from commit 381ecb8de19251a146249a49784c2a6c779f6676)
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/5011
Reviewed-by: Juan Yu <jyu@cloudera.com>