Randomize the order scan nodes process files.
In very reasonable cases, having all the scan nodes read files in the
same order (different blocks of a single HDFS file), can cause uneven
load on the DN that need to serve remote reads. This could help with
the transceiver limit issues we are seeing but seems sensible to do
regardless.

Change-Id: I85f18b97cfe991bbb246eac2291c3d57d8248855
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/5034
Tested-by: jenkins
Reviewed-by: Nong Li <nong@cloudera.com>