IMPALA-1444 Using both ldaps:// and --ldap_tls breaks LDAP auth
There are two ways to get TLS on the LDAP connection.  The first is to
use an ldaps:// URI. The second is to use an ldap:// URI, but then
issue a STARTTLS command to upgrade the insecure connection to a
secure connection.  This is what the -ldap_tls argument does.

Supplying both of these together doesn't make sense, and confuses the
LDAP server.  The fix is to simply ignore -ldap_tls when ldaps:// is
used.

Change-Id: Id9238c24324414d4d00090bb3e926f217a4be8aa
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/5075
Reviewed-by: Henry Robinson <henry@cloudera.com>
Tested-by: jenkins