IMPALA-1442: Better fix for non-buffered SASL transports
The Thrift SASL implementation relies on the same transport being used
going in and out for the same peer socket (see
TSaslServerTransport::Factory::getTransport()). This means that the
transport's factory can't be used compositionally, as intended in the
last patch for this issue, as the wrapper factory has to be sure to hand
the SASL factory the same transport every time the same socket is passed
in, but we were creating a new buffered transport on each occasion.

This patch does something much simpler: the SASL factory is forced to
always create a buffered transport. This rather breaks compositionality,
but it was already broken by the factory's implementation.

Change-Id: Iec1a4337778e774ab8748e0412515abd07399c94
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/5163
Reviewed-by: Matthew Jacobs <mj@cloudera.com>
Tested-by: jenkins