Fix zlib maximum compressed size estimate
Some compressed file formats in Impala, such as SEQUENCEFILE and TEXT, have run into
errors with zlib compression. This has been tracked to situations where the output
buffer is too small because of an error in zlib.

There is a known issue that zlib 1.2.3 does not include the size of the
gzip wrapper when estimating the size of compressed output.
This is has been fixed in zlib 1.2.3.1: http://www.zlib.net/ChangeLog.txt

"Take into account wrapper variations in deflateBound()"

Mark, maintainer of zlib, has stated that 12 needs to be added to result for gzip,
http://compgroups.net/comp.unix.programmer/gzip-compressing-an-in-memory-string-usi/54854

To have a safe upper bound for "wrapper variations", we add 32 to estimate.

Change-Id: I4f17506e880d2d1cdeca8e770434efb202be69ff
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/5249
Reviewed-by: Skye Wanderman-Milne <skye@cloudera.com>
Tested-by: jenkins