Parametrized Test Execution
The goal of this patch is to allow configuring which tests to run during
the build. The options added in run-all-tests.sh are added to be
generated from Jenkins. If Jenkins is not setting these options, all
tests are executed.

An environment that would only run the backend tests would look like the
following:

FE=false EE_TEST=false CLUSTER_TEST=false JDBC_TEST=false \
./bin/run-all-tests.sh

In addition, we enable generating JUnit compatible XML files for the
backend tests for easier tracing in CI builds.

Change-Id: Ibf257d37757aecbb75c9194345e494cacdcdc326
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/5261
Reviewed-by: Martin Grund <mgrund@cloudera.com>
Tested-by: jenkins