Improve minimum CPU requirement checks
Always try to execute an SSSE3 instruction.  Otherwise, we may continue
to run if we get "lucky" and the compiler doesn't emit an SSSE3
instruction in this particular build which is often the case in debug
builds, but also happens every now and then in release builds.  This has
been a big source of confusion as to what Impala's minimum CPU
requirements actually are (e.g. in some cases, Impala 1.4 would run on
pre-SSSE3 CPUs, but Impala 2.0 does not, even though neither supported
those CPUs).  Also adjust the error message to indicate we are likely
going to "exit" now because CPU requirements aren't met.

Tested the log-flush-then-crash logic using ud2 instruction and verified
the message appears in the log file (since I don't have access to a
pre-SSSE3 cpu).

Change-Id: I0413cddf41e8a2b4789f5c6a3bc0d9e6c18b5767
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/5310
Reviewed-by: Daniel Hecht <dhecht@cloudera.com>
Tested-by: jenkins