IMPALA-1479: Downgrade severity of unknown query ID for report
The error message that was printed at INFO when a report is received for
an unknown query isn't really actionable, and can rapidly dominate a log
if there are many backends that are slow to tear down once a query is
finished. Any underlying problems with fragments continuing to run after
the query is finished should be detected through other means (e.g. the
debug webpage).

Change-Id: I156389f42a3fada1318020d4731c3da64793f7b1
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/5378
Reviewed-by: Nong Li <nong@cloudera.com>
Reviewed-by: Henry Robinson <henry@cloudera.com>
Tested-by: Henry Robinson <henry@cloudera.com>