Fix HdfsLzoTextScanner::Close()
The LZO scanner did not call HdfsTextScanner::Close(), instead
replicating the logic in its own Close() method. This makes it harder
to keep in sync with the text scanner, such as in this case, where it
was not closing the new boundary_pool_. This patch has the LZO scanner
call HdfsTextScanner::Close() instead of replicating it.

Change-Id: Ifed531e67abc4824a6489b9dfcd2d55d7b50547a
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/5395
Reviewed-by: Matthew Jacobs <mj@cloudera.com>
Tested-by: Skye Wanderman-Milne <skye@cloudera.com>