Remove data compaction flag from BE. Also, some simple readability improvements.
The purpose of the data compaction flag was to have compact row batches
when constructing hash tables during join builds (e.g., such that
the hash table would not have to hang on to I/O buffers).
The FE used to mark the buid-side scans as compacting.

The new partitioned join must copy all string data anyway,
so there is no more need for compaction on the scan side.

The obsolete compaction is not set in the FE anymore, and this
patch removes it from the BE also.

Change-Id: Ie2208f44fda33f0bafee61559ac29e26977cfe4e
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/5494
Reviewed-by: Daniel Hecht <dhecht@cloudera.com>
Tested-by: jenkins