IMPALA-1808: AnalyticEvalNode cannot handle partition/order by exprs with NaN
Analytic function evaluation was broken when partition or order by exprs
evaluated to NaN (i.e. 0/0). We were relying on the comparison of the
current row with the previous row to be equal (i.e. x == x), but x != x
if x is NaN, and in the case of the very first row in the stream, some
logic breaks if x != x. The fix is to handle the very first row specially.

Change-Id: I1c33445d55a70c7f107f05eeadef272b7973ee11
Reviewed-on: http://gerrit.cloudera.org:8080/179
Reviewed-by: Matthew Jacobs <mj@cloudera.com>
Tested-by: Internal Jenkins